Vectors to tuples, doubles to ints

+++ src/Farm.cpp
@@ -5,7 +5,7 @@
 // #include "Slaughter_transport.h"
 
 //Farm::Farm(int in_id, int in_x, int in_y, short in_size)
-Farm::Farm(double in_id, std::string in_fips, double in_x, double in_y, double in_size)
+Farm::Farm(int in_id, std::string in_fips, double in_x, double in_y, int in_size)
 {
 	id = in_id;
 	fips = in_fips;
@@ -50,7 +50,7 @@ Farm::~Farm()
 // }
 // 
 
-void Farm::set_cellID(const double in_cellID)
+void Farm::set_cellID(const int in_cellID)
 {
 	cellID = in_cellID;
 }
